shared q3config is now only for the default mod and is enabled by default
It's to avoid config overwriting caused by mods like ratmod or aftershock that use initialized cvars. OmegA doesn't use it so we can share the config between it and OpenArena. If you have ratmod as your default mod it's probably that you don't mind having simple items and stuff enabled in baseoa too.